+++ awe-framework/awe-modules/awe-rest/src/main/java/com/almis/awe/rest/dto/RequestParameter.java
@@ -2,10 +2,14 @@ package com.almis.awe.rest.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import lombok.experimental.Accessors;
 
+import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 @Data
+@Accessors(chain = true)
 public class RequestParameter {
 
   @Schema(title = "Parameters map in Json format",