+++ awe-framework/awe-modules/awe-rest/src/main/java/com/almis/awe/rest/service/JWTTokenService.java
@@ -1,6 +1,7 @@
 package com.almis.awe.rest.service;
 
 import com.almis.awe.config.ServiceConfig;
+import com.almis.awe.rest.dto.JwtTokenInfo;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
@@ -62,19 +63,28 @@ public class JWTTokenService extends ServiceConfig {
    * @param authentication Authentication object
    * @param response {@link HttpServletResponse}
    */
-  public void generateToken(Authentication authentication, HttpServletResponse response) {
+  public JwtTokenInfo generateToken(Authentication authentication, HttpServletResponse response) {
+    final String username = ((UserDetails) authentication.getPrincipal()).getUsername();
+    final Date expiresAt = new Date(System.currentTimeMillis() + expiration.toMillis());
+
     String token = JWT.create()
-            .withSubject(((UserDetails) authentication.getPrincipal()).getUsername())
-            .withExpiresAt(new Date(System.currentTimeMillis() + expiration.toMillis()))
+            .withSubject(username)
+            .withExpiresAt(expiresAt)
             .withIssuer(issuer)
             .sign(Algorithm.HMAC512(secret.getBytes()));
 
     // Add auth header
     response.addHeader(authorizationHeader, prefix + " " + token);
+
+    return new JwtTokenInfo()
+      .setToken(token)
+      .setIssuer(issuer)
+      .setUsername(username)
+      .setExpiredAt(expiresAt);
   }
 
   /**
-   * Verify token. Check if it has JWT format and if not expiredh
+   * Verify token. Check if it has JWT format and if not expired
    * @param token JWT token
    * @return Decode JWT token
    */