+++ awe-framework/awe-modules/awe-scheduler/src/main/java/com/almis/awe/scheduler/service/report/ISchedulerReportService.java
@@ -1,0 +1,13 @@
+package com.almis.awe.scheduler.service.report;
+
+import com.almis.awe.scheduler.bean.task.Task;
+import com.almis.awe.scheduler.bean.task.TaskExecution;
+import com.almis.awe.scheduler.enums.ReportType;
+
+public interface ISchedulerReportService {
+  default ReportType getType() {
+    return null;
+  }
+
+  void execute(Task task, TaskExecution execution);
+}