+++ awe-tests/awe-boot/src/test/java/com/almis/awe/test/integration/controller/ApiRestControllerTest.java
@@ -2,6 +2,7 @@ package com.almis.awe.test.integration.controller;
 
 import com.almis.awe.model.type.AnswerType;
 import com.almis.awe.rest.dto.AweRestResponse;
+import com.almis.awe.rest.dto.LoginRequest;
 import com.almis.awe.rest.dto.LoginResponse;
 import com.almis.awe.rest.dto.RequestParameter;
 import com.almis.awe.rest.service.JWTTokenService;
@@ -62,7 +63,7 @@ class ApiRestControllerTest extends AbstractSpringFixedEnvironmentIT {
       .withExpiresAt(new Date(System.currentTimeMillis() + 60000)) // 1 min
       .withIssuer("AWE ISSUER")
       .sign(Algorithm.HMAC512(jwtTokenService.getSecret().getBytes()));
-    headers.add("Authorization", jwtTokenService.getPrefix() + " " + jwtToken);
+    headers.setBearerAuth(jwtToken);
   }
 
   @Nested
@@ -73,12 +74,14 @@ class ApiRestControllerTest extends AbstractSpringFixedEnvironmentIT {
     @Test
     void authenticate_success() {
       // Build entity and variable maps
-      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"))
-              .queryParam("username", "test")
-              .queryParam("password", "test");
-      HttpEntity<String> entity = new HttpEntity<>(headers);
+      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"));
+      LoginRequest loginRequest = new LoginRequest()
+        .setUsername("test")
+        .setPassword("test");
+      HttpEntity<LoginRequest> loginRequestEntity = new HttpEntity<>(loginRequest, headers);
+
       ResponseEntity<LoginResponse> response = restTemplate.exchange(builder.toUriString(),
-              HttpMethod.POST, entity, LoginResponse.class);
+              HttpMethod.POST, loginRequestEntity, LoginResponse.class);
       assertEquals(HttpStatus.OK.value(), response.getStatusCodeValue());
       assertNotNull(Objects.requireNonNull(response.getBody()).getToken());
     }
@@ -86,12 +89,13 @@ class ApiRestControllerTest extends AbstractSpringFixedEnvironmentIT {
     @Test
     void authenticateWithBadCredentials_unauthorized() {
       // Build entity and variable maps
-      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"))
-              .queryParam("username", "test")
-              .queryParam("password", "dummy");
-      HttpEntity<String> entity = new HttpEntity<>(headers);
+      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"));
+      LoginRequest loginRequest = new LoginRequest()
+        .setUsername("test")
+        .setPassword("dummy");
+      HttpEntity<LoginRequest> loginRequestEntity = new HttpEntity<>(loginRequest, headers);
       ResponseEntity<AweRestResponse> response = restTemplate.exchange(builder.toUriString(),
-              HttpMethod.POST, entity, AweRestResponse.class);
+              HttpMethod.POST, loginRequestEntity, AweRestResponse.class);
       assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCodeValue());
       assertEquals("Bad credentials", Objects.requireNonNull(response.getBody()).getMessage());
     }
@@ -99,12 +103,13 @@ class ApiRestControllerTest extends AbstractSpringFixedEnvironmentIT {
     @Test
     void authenticateWithUserNotFound_unauthorized() {
       // Build entity and variable maps
-      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"))
-              .queryParam("username", "foo")
-              .queryParam("password", "dummy");
-      HttpEntity<String> entity = new HttpEntity<>(headers);
+      UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(createURLWithPort("/api/authenticate"));
+      LoginRequest loginRequest = new LoginRequest()
+        .setUsername("foo")
+        .setPassword("dummy");
+      HttpEntity<LoginRequest> loginRequestEntity = new HttpEntity<>(loginRequest, headers);
       ResponseEntity<AweRestResponse> response = restTemplate.exchange(builder.toUriString(),
-              HttpMethod.POST, entity, AweRestResponse.class);
+              HttpMethod.POST, loginRequestEntity, AweRestResponse.class);
       assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCodeValue());
       assertEquals("User not found or not active", Objects.requireNonNull(response.getBody()).getMessage());
     }